pipelines/postgres-nfl-tandl/pyproject.toml
[tool.poetry]
name = "postgres-nfl-tandl"
version = "0.1.0"
description = ""
authors = ["Christopher Mortimer <christopher@mortimer.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beammeup = { path = "../../lib/beammeup", develop = true }
python-dotenv = "^1.0.1"
hydra-core = "^1.3.2"
click = "^8.1.7"
rich = "^13.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
